--- eroom/src/main/resources/templates/admin.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>admin</title>
-</head>
-<body>
-
-</body>
-</html>
