+++ eroom/src/main/resources/templates/checkinout.html
@@ -7,7 +7,7 @@
   layout:decorate="~{/include/layout}">
 <head>
   <meta charset="UTF-8">
-  <title>출퇴근 카드</title>
+  <title>Home</title>
   <style>
     .attendance-card {
       max-width: 380px;
@@ -44,21 +44,21 @@
 <body>
 <main class="main" id="top" layout:fragment="content">
 
-  <div class="attendance-card">
+<!--   <div class="attendance-card">
     <div class="card shadow-sm border-0">
       <div class="card-body text-center">
 		<form id="attendanceForm" action="/attendance/log" method="post">
-			<!-- csrf 토큰 -->
+			csrf 토큰
 			<input type="hidden" th:name="${_csrf.parameterName}" th:value="${_csrf_token}">
-			<!-- 로그인한 사용자 -->
+			로그인한 사용자
 			<input type="hidden" th:value="${#authentication.principal.employee.employeeNo}" name="employeeNo">
-			<!-- 출근 / 퇴근 type hidden -->
+			출근 / 퇴근 type hidden
 			<input type="hidden" name="attendanceType" id="attendanceType">
 			
 			
-	        <!-- 현재 시간 -->
+	        현재 시간
 	        <div class="current-time fw-semibold text-secondary fs-8 mb-0" id="currentTime"></div><br>
-	        <!-- 프로필 + 이름 / 부서 + 직급 -->
+	        프로필 + 이름 / 부서 + 직급
 	        <div class="d-flex justify-content-center mb-3">
 	          <div class="d-flex align-items-center">
 	            <img id="profileImage" class="rounded-circle me-3 profile-img" src="/assets/img/team/15.webp" alt="프로필">
@@ -68,7 +68,7 @@
 	            </div>
 	          </div>
 	        </div>
-	        <!-- 출퇴근 버튼 -->
+	        출퇴근 버튼
 	        <div class="btn-wrap d-flex justify-content-center gap-2">
 			  	        
 	          <button id="checkInBtn" class="btn btn-outline-primary btn-sm">출근하기</button>
@@ -196,7 +196,7 @@
     		console.error('퇴근 오류',error);
     		alert('퇴근 실패');
     	});
-    });
+    }); -->
     
   /*   // 페이지가 로드 되면 상태 복원, 초기화
     document.addEventListener('DOMContentLoaded',function(){
