+++ eroom/src/main/resources/templates/sample.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html 
+lang="en-US" dir="ltr" data-navigation-type="default" data-navbar-horizontal-shape="default"
+	  xmlns:th="http://www.thymeleaf.org"
+	  xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+	  layout:decorate="~{/include/layout}">
+  <body>
+    <main class="main" id="top" layout:fragment="content">
+    
+    </main>
+  </body>
+</html>
